ml-models/requirements.txt
fastapi
numpy
opencv_python
Pillow
tensorflow
uvicorn
python-multipart
os